Implemented WorkshopManager VesselModule and moved GCScenario to use it.
WorkshopManager collects GroundWorkshop modules and makes a list of
WorkshopInfos, providing a UI for GCScenario and replacing the
WorkshopVesselInfo framework.

This is more robust as VesselModule is created and initialized with the game and
exists in both flight and KSC/Map/etc, in loaded and unloaded vessels. It also
saves its own WorkshopInfo list, freeing scenario module of storing data.

Also added celestial body tabs that group workshops located on the same CB to
unclutter the workshop list. Closed #13