Added InSituAssemblySpace and GroundAssemblyWorkshop modules.
InSituAssemblySpace uses the same space as the DeployableKitContainer in the
same part. This allows the use of empty kit containers as assembly spaces that
will store the assembled kit themselves.
GroundAssemblyLine is a complete analog to the GroundWorkshop, but for assembly.

DIYKit part uses InSituAssemblySpace.
GroundAssemblyLine part uses GroundAssemblyLine MODULE.

Added IContainerProducer interface, which is implemented by the AssemblySpace
module. So now the AssemblySpace (via AssemblyWorkshop's UI) can instantly spawn
an empty kit to be used as a one-time assembly space for something big.