I35 working launch-and-dock algorithm for symmetrict vessels
This algorithm only docks the launched vessel if a free attach node node is
found that is both collinear with the recipient attach node and is very close to
it. Which means that only a symmetric vessel that "ends" with an attach mode
along its axis can be docked. Which is bad for station construction.