Implemented initial vessel/part recycling logic.
Vessels are recycled from leaf-parts to the root in depth-first order.

For all vessels, parts and all their parents are skipped if: they're crewed;
their resources cannot be transferred away.

For current vessel the rootPart and the whole branch from the rootPart to the
current part is skipped.

Parts are recycled into AssemblyResource and ConstructionResource according to
their PartKit designs, with maximum efficiency capped by Global constants and
the skill level of the most skilled engineer inside the workshop. Recycling also
requires energy.